fix: page.tsx 서버컴포넌트로변경, ui폴더삭제
